@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<TargetFramework>netstandard2.0</TargetFramework>
-		<LangVersion>9.0</LangVersion>
+		<LangVersion>10.0</LangVersion>
 
 		<!-- Special properties for analyzer packages. -->
 		<IncludeBuildOutput>false</IncludeBuildOutput>
@@ -15,6 +15,13 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.10.0" />
 	</ItemGroup>
+	<ItemGroup>
+		<Using Include="System" />
+		<Using Include="System.Collections.Generic" />
+		<Using Include="System.Collections.Concurrent" />
+		<Using Include="System.Linq" />
+		<Using Include="System.IO" />
+	</ItemGroup>
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.10.0" PrivateAssets="all" GeneratePathProperty="true" />
@@ -22,7 +29,7 @@
 		<None Include="$(OutputPath)$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
 	</ItemGroup>
 
-   <PropertyGroup>
+   <!--<PropertyGroup>
         <GetTargetPathDependsOn>$(GetTargetPathDependsOn);BlubbiChanged-GetDependencyTargetPaths</GetTargetPathDependsOn>
     </PropertyGroup>
 
@@ -30,5 +37,5 @@
         <ItemGroup>
             <TargetPathWithTargetPlatformMoniker Include="$(PKGMicrosoft_CodeAnalysis_CSharp)\lib\netstandard2.0\Microsoft.CodeAnalysis.CSharp.dll" IncludeRuntimeDependency="true" />
         </ItemGroup>
-    </Target>
+    </Target>-->
 </Project>
