@@ -1,10 +1,9 @@
-﻿namespace AppBrokerASP.Configuration
-{
-    public class PainlessMeshSettings
-    {
-        public const string ConfigName = nameof(PainlessMeshSettings);
+﻿namespace AppBrokerASP.Configuration;
 
-        public bool Enabled { get; set; }
-        public ushort ListenPort { get; set; }
-    }
+public class PainlessMeshSettings
+{
+    public const string ConfigName = nameof(PainlessMeshSettings);
+
+    public bool Enabled { get; set; }
+    public ushort ListenPort { get; set; }
 }
