@@ -1,12 +1,11 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AppBrokerASP.Database.Model
+namespace AppBrokerASP.Database.Model;
+
+public class DeviceModel
 {
-    public class DeviceModel 
-    {
-        [Key]
-        public long Id { get; set; }
-        public string TypeName { get; set; } = "";
-        public string FriendlyName { get; set; } = "";
-    }
+    [Key]
+    public long Id { get; set; }
+    public string TypeName { get; set; } = "";
+    public string FriendlyName { get; set; } = "";
 }
