@@ -1,13 +1,1 @@
-﻿namespace AppBrokerASP.Database.Model
-{
-    //public class HeaterConfigTemplateModel
-    //{
-    //    [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    //    public long Id { get; set; }
-
-    //    public string TemplateName { get; set; }
-
-    //    [ForeignKey("DeviceId")]
-    //    public virtual DeviceModel Device { get; set; }
-    //}
-}
+﻿namespace AppBrokerASP.Database.Model;
