@@ -15,319 +15,327 @@ using System.Net.Http;
 using SocketIOClient;
 using AppBrokerASP.Extension;
 using Newtonsoft.Json.Linq;
+using AppBroker.Core;
+using AppBroker.Core.Devices;
 
-namespace AppBrokerASP
+namespace AppBrokerASP;
+
+public class DeviceManager : IDisposable, IDeviceManager
 {
-    public class DeviceManager : IDisposable
+    public ZigbeeConfig Config { get; }
+    public ConcurrentDictionary<long, Device> Devices { get; } = new();
+    public IReadOnlyCollection<Type> DeviceTypes => types;
+
+    public event EventHandler<(long id, Device device)>? NewDeviceAdded;
+
+    private SocketIO? client;
+    private bool disposed;
+    private readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
+    private readonly List<Type> types;
+    private readonly Dictionary<string, Type> alternativeNamesForTypes = new(StringComparer.OrdinalIgnoreCase);
+    readonly Task temp;
+    private static readonly HttpClient http = new();
+
+    public DeviceManager()
     {
-        public ZigbeeConfig Config { get; }
-        public ConcurrentDictionary<long, Device> Devices = new();
-        private SocketIO? client;
-        private bool disposed;
-        private readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
-        private readonly List<Type> types;
-        private readonly Dictionary<string, Type> alternativeNamesForTypes = new(StringComparer.OrdinalIgnoreCase);
-        readonly Task temp;
-        private static readonly HttpClient http = new();
+        Config = InstanceContainer.Instance.ConfigManager.ZigbeeConfig;
 
-        public DeviceManager()
+        types = Assembly
+            .GetExecutingAssembly()
+            .GetTypes()
+            .Where(x => typeof(Device).IsAssignableFrom(x) && x != typeof(Device))
+            .ToList();
+
+        foreach (var type in types)
         {
-            Config = InstanceContainer.ConfigManager.ZigbeeConfig;
-
-            types = Assembly
-                .GetExecutingAssembly()
-                .GetTypes()
-                .Where(x => typeof(Device).IsAssignableFrom(x) && x != typeof(Device))
-                .ToList();
-
-            foreach (var type in types)
+            var names = GetAllNamesFor(type);
+            foreach (var name in names)
             {
-                var names = GetAllNamesFor(type);
-                foreach (var name in names)
+                alternativeNamesForTypes[name] = type;
+            }
+        }
+
+        InstanceContainer.Instance.MeshManager.NewConnectionEstablished += Node_NewConnectionEstablished;
+        InstanceContainer.Instance.MeshManager.ConnectionLost += MeshManager_ConnectionLost;
+        InstanceContainer.Instance.MeshManager.ConnectionReastablished += MeshManager_ConnectionReastablished;
+
+        temp = ConnectToIOBroker();
+    }
+    private void AddNewDeviceToDic(long id, Device device)
+    {
+        if (Devices.TryAdd(id, device))
+            NewDeviceAdded?.Invoke(this, (id, device));
+    }
+
+    private static List<string> GetAllNamesFor(Type y)
+    {
+        var names = new List<string>();
+        var attribute = y.GetCustomAttribute<DeviceNameAttribute>();
+
+        if (attribute is not null)
+        {
+            names.Add(attribute.PreferredName);
+            names.AddRange(attribute.AlternativeNames);
+        }
+        names.Add(y.Name);
+        return names;
+    }
+
+    private void MeshManager_ConnectionLost(object? sender, uint e)
+    {
+        if (Devices.TryGetValue(e, out var device))
+        {
+            device.StopDevice();
+        }
+    }
+    private void MeshManager_ConnectionReastablished(object? sender, (uint id, ByteLengthList parameter) e)
+    {
+        if (Devices.TryGetValue(e.id, out var device))
+        {
+            device.Reconnect(e.parameter);
+        }
+    }
+
+    private void Node_NewConnectionEstablished(object? sender, (Sub c, ByteLengthList l) e)
+    {
+        if (Devices.TryGetValue(e.c.NodeId, out var device))
+        {
+            device.Reconnect(e.l);
+        }
+        else
+        {
+            var deviceName = Encoding.UTF8.GetString(e.l[1]);
+            var newDevice = CreateDeviceFromName(deviceName, e.c.NodeId, e.l);
+
+            if (newDevice is null)
+                return;
+
+            logger.Debug($"New Device: {newDevice.TypeName}, {newDevice.Id}");
+            AddNewDeviceToDic(e.c.NodeId, newDevice);
+            //_ = Devices.TryAdd(e.c.NodeId, newDevice);
+
+            if (!DbProvider.AddDeviceToDb(newDevice))
+                _ = DbProvider.MergeDeviceWithDbData(newDevice);
+        }
+    }
+
+    private Device? CreateDeviceFromName(string deviceName, params object[] ctorArgs)
+    {
+        logger.Trace($"Trying to get device with {deviceName} name");
+
+        if (!alternativeNamesForTypes.TryGetValue(deviceName, out var type) || type is null)
+        {
+            logger.Error($"Failed to get device with name {deviceName}");
+            return null;
+        }
+
+        var newDeviceObj = Activator.CreateInstance(type, ctorArgs);
+
+        if (newDeviceObj is null || newDeviceObj is not Device newDevice)
+        {
+            logger.Error($"Failed to get create device {deviceName}");
+            return null;
+        }
+
+        return newDevice;
+    }
+
+    private async Task ConnectToIOBroker()
+    {
+        client = new SocketIO(new Uri(Config.SocketIOUrl), new SocketIOOptions()
+        {
+            EIO = Config.NewSocketIoversion ? 4 : 3
+        });
+
+        client.OnAny(async (eventName, response) =>
+        {
+            if (IoBrokerZigbee.TryParse(eventName, response.ToString(), out var zo) && zo is not null)
+            {
+                if (Devices.TryGetValue(zo.Id, out var dev) && dev is ZigbeeDevice zigbeeDev)
                 {
-                    alternativeNamesForTypes[name] = type;
-                }
-            }
-
-            InstanceContainer.MeshManager.NewConnectionEstablished += Node_NewConnectionEstablished;
-            InstanceContainer.MeshManager.ConnectionLost += MeshManager_ConnectionLost;
-            InstanceContainer.MeshManager.ConnectionReastablished += MeshManager_ConnectionReastablished;
-
-            temp = ConnectToIOBroker();
-        }
-
-        private static List<string> GetAllNamesFor(Type y)
-        {
-            var names = new List<string>();
-            var attribute = y.GetCustomAttribute<DeviceNameAttribute>();
-
-            if (attribute is not null)
-            {
-                names.Add(attribute.PreferredName);
-                names.AddRange(attribute.AlternativeNames);
-            }
-            names.Add(y.Name);
-            return names;
-        }
-
-        private void MeshManager_ConnectionLost(object? sender, uint e)
-        {
-            if (Devices.TryGetValue(e, out var device))
-            {
-                device.StopDevice();
-            }
-        }
-        private void MeshManager_ConnectionReastablished(object? sender, (uint id, ByteLengthList parameter) e)
-        {
-            if (Devices.TryGetValue(e.id, out var device))
-            {
-                device.Reconnect(e.parameter);
-            }
-        }
-
-        private void Node_NewConnectionEstablished(object? sender, (Sub c, ByteLengthList l) e)
-        {
-            if (Devices.TryGetValue(e.c.NodeId, out var device))
-            {
-                device.Reconnect(e.l);
-            }
-            else
-            {
-                var deviceName = Encoding.UTF8.GetString(e.l[1]);
-                var newDevice = CreateDeviceFromName(deviceName, e.c.NodeId, e.l);
-
-                if (newDevice is null)
-                    return;
-
-                logger.Debug($"New Device: {newDevice.TypeName}, {newDevice.Id}");
-                _ = Devices.TryAdd(e.c.NodeId, newDevice);
-
-                if (!DbProvider.AddDeviceToDb(newDevice))
-                    _ = DbProvider.MergeDeviceWithDbData(newDevice);
-            }
-        }
-
-        private Device? CreateDeviceFromName(string deviceName, params object[] ctorArgs)
-        {
-            logger.Trace($"Trying to get device with {deviceName} name");
-
-            if (!alternativeNamesForTypes.TryGetValue(deviceName, out var type) || type is null)
-            {
-                logger.Error($"Failed to get device with name {deviceName}");
-                return null;
-            }
-
-            var newDeviceObj = Activator.CreateInstance(type, ctorArgs);
-
-            if (newDeviceObj is null || newDeviceObj is not Device newDevice)
-            {
-                logger.Error($"Failed to get create device {deviceName}");
-                return null;
-            }
-
-            return newDevice;
-        }
-
-        private async Task ConnectToIOBroker()
-        {
-            client = new SocketIO(new Uri(Config.SocketIOUrl), new SocketIOOptions()
-            {
-                EIO = Config.NewSocketIoversion ? 4 : 3
-            });
-
-            client.OnAny(async (eventName, response) =>
-            {
-                if (IoBrokerZigbee.TryParse(eventName, response.ToString(), out var zo) && zo is not null)
-                {
-                    if (Devices.TryGetValue(zo.Id, out var dev) && dev is ZigbeeDevice zigbeeDev)
+                    try
                     {
-                        try
-                        {
-                            zigbeeDev.SetPropFromIoBroker(zo, true);
-                        }
-                        catch (Exception e)
-                        {
-                            logger.Error(e);
-                        }
+                        zigbeeDev.SetPropFromIoBroker(zo, true);
                     }
-                    else
+                    catch (Exception e)
                     {
-                        try
-                        {
-
-                            await GetZigbeeDevices(client);
-                        }
-                        catch (Exception ex)
-                        {
-                            logger.Error(ex);
-                        }
+                        logger.Error(e);
                     }
                 }
-            });
-
-            client.OnError += async (sender, args) =>
-            {
-                logger.Error($"AfterException, trying reconnect: {args}");
-                await client.DisconnectAsync();
-                await client.ConnectAsync();
-            };
-
-            client.OnConnected += async (s, e) =>
-            {
-                logger.Debug("Connected Zigbee Client");
-                await client.EmitAsync("subscribe", "zigbee.*");
-                await client.EmitAsync("subscribeObjects", '*');
-                try
+                else
                 {
+                    try
+                    {
 
-                    await GetZigbeeDevices(client);
+                        await GetZigbeeDevices(client);
+                    }
+                    catch (Exception ex)
+                    {
+                        logger.Error(ex);
+                    }
                 }
-                catch (Exception ex)
-                {
-                    logger.Error(ex);
-                }
-            };
+            }
+        });
 
+        client.OnError += async (sender, args) =>
+        {
+            logger.Error($"AfterException, trying reconnect: {args}");
+            await client.DisconnectAsync();
             await client.ConnectAsync();
+        };
+
+        client.OnConnected += async (s, e) =>
+        {
+            logger.Debug("Connected Zigbee Client");
+            await client.EmitAsync("subscribe", "zigbee.*");
+            await client.EmitAsync("subscribeObjects", '*');
+            try
+            {
+
+                await GetZigbeeDevices(client);
+            }
+            catch (Exception ex)
+            {
+                logger.Error(ex);
+            }
+        };
+
+        await client.ConnectAsync();
+    }
+
+    public async Task GetZigbeeDevices(SocketIO socket)
+    {
+        if (client is null)
+        {
+            logger.Error("SocketIO Client is null");
+            return;
         }
 
-        public async Task GetZigbeeDevices(SocketIO socket)
+        var allObjectsResponse = await socket.Emit("getObjects");
+        var allObjectscontentNew = allObjectsResponse?.GetValue(1).ToString();
+
+        if (allObjectscontentNew is null)
         {
-            if (!Devices.IsEmpty)
+            logger.Error("getObjects response is empty");
+            return;
+        }
+
+        var ioBrokerObject = JsonConvert.DeserializeObject<Dictionary<string, JToken>>(allObjectscontentNew)?
+            .Where(x => x.Key.StartsWith("zigbee.0."))!
+            .ToDictionary(x => x.Key, x => x.Value.ToObject<ZigbeeIOBrokerProperty>())
+            .Values.AsEnumerable();
+
+        if (ioBrokerObject is null)
+        {
+            logger.Error($"Error deserializing IOBroker Property");
+
+            return;
+        }
+        var idRequest = @$"{Config.HttpUrl}/get/";
+        var idsAlreadyInRequest = new List<ulong>();
+        var stateRequest = @$"{Config.HttpUrl}/get/";
+        foreach (var item in ioBrokerObject)
+        {
+            if (item is null)
+                continue;
+
+            var matches = item._id.Split('.');
+            if (matches.Length > 2)
             {
-                logger.Debug($"Cancel {nameof(GetZigbeeDevices)}, because it wasn't the startup");
-                return;
+                if (ulong.TryParse(matches[2].ToString(), System.Globalization.NumberStyles.HexNumber, null, out var id))
+                {
+                    if (!idsAlreadyInRequest.Contains(id))
+                    {
+                        idRequest += string.Join('.', matches.Take(3)) + ",";
+                        idsAlreadyInRequest.Add(id);
+                    }
+                    stateRequest += item._id + ",";
+                }
             }
+        }
 
-            if (client is null)
+        var content = await http.GetStringAsync(idRequest);
+
+        var getDeviceResponses = JsonConvert.DeserializeObject<IoBrokerGetDeviceResponse[]>(content)!;
+        content = await http.GetStringAsync(stateRequest);
+        var deviceStates = JsonConvert.DeserializeObject<IoBrokerStateResponse[]>(content)!;
+
+        foreach (var deviceRes in getDeviceResponses)
+        {
+            if (!long.TryParse(deviceRes.native.id, System.Globalization.NumberStyles.HexNumber, null, out var id))
+                continue;
+
+            if (!Devices.TryGetValue(id, out var dev))
             {
-                logger.Error("SocketIO Client is null");
-                return;
-            }
+                dev = CreateDeviceFromName(deviceRes.common.type, id, client);
 
-            var allObjectsResponse = await socket.Emit("getObjects");
-            var allObjectscontentNew = allObjectsResponse?.GetValue(1).ToString();
-
-            if (allObjectscontentNew is null)
-            {
-                logger.Error("getObjects response is empty");
-                return;
-            }
-
-            var ioBrokerObject = JsonConvert.DeserializeObject<Dictionary<string, JToken>>(allObjectscontentNew)?
-                .Where(x => x.Key.StartsWith("zigbee.0."))!
-                .ToDictionary(x => x.Key, x => x.Value.ToObject<ZigbeeIOBrokerProperty>())
-                .Values.AsEnumerable();
-
-            if (ioBrokerObject is null)
-            {
-                logger.Error($"Error deserializing IOBroker Property");
-
-                return;
-            }
-            var idRequest = @$"{Config.HttpUrl}/get/";
-            var idsAlreadyInRequest = new List<ulong>();
-            var stateRequest = @$"{Config.HttpUrl}/get/";
-            foreach (var item in ioBrokerObject)
-            {
-                if (item is null)
+                if (dev is null or default(Device))
+                {
+                    logger.Warn($"Found not mapped device: {deviceRes.common.name} ({deviceRes.common.type})");
                     continue;
-
-                var matches = item._id.Split('.');
-                if (matches.Length > 2)
-                {
-                    if (ulong.TryParse(matches[2].ToString(), System.Globalization.NumberStyles.HexNumber, null, out var id))
-                    {
-                        if (!idsAlreadyInRequest.Contains(id))
-                        {
-                            idRequest += string.Join('.', matches.Take(3)) + ",";
-                            idsAlreadyInRequest.Add(id);
-                        }
-                        stateRequest += item._id + ",";
-                    }
                 }
+
+                if (dev is ZigbeeDevice zd)
+                    zd.AdapterWithId = deviceRes._id;
+
+                //_ = Devices.TryAdd(id, dev);
+
+                if (!DbProvider.AddDeviceToDb(dev))
+                    _ = DbProvider.MergeDeviceWithDbData(dev);
+
+                // If name is only numbers, try to get better name
+                if (string.IsNullOrWhiteSpace(dev.FriendlyName) || long.TryParse(dev.FriendlyName, out _))
+                    dev.FriendlyName = deviceRes.common.name;
+
+                AddNewDeviceToDic(id, dev);
             }
 
-            var content = await http.GetStringAsync(idRequest);
-
-            var getDeviceResponses = JsonConvert.DeserializeObject<IoBrokerGetDeviceResponse[]>(content)!;
-            content = await http.GetStringAsync(stateRequest);
-            var deviceStates = JsonConvert.DeserializeObject<IoBrokerStateResponse[]>(content)!;
-
-            foreach (var deviceRes in getDeviceResponses)
+            foreach (var item in deviceStates.Where(x => x._id.Contains(deviceRes.native.id)))
             {
-                if (!long.TryParse(deviceRes.native.id, System.Globalization.NumberStyles.HexNumber, null, out var id))
-                    continue;
-
-                if (!Devices.TryGetValue(id, out var dev))
+                var ioObject = new IoBrokerObject(BrokerEvent.StateChange, "", 0, item.common.name.ToLower().Replace(" ", "_"), new Parameter(item.val));
+                if (dev is XiaomiTempSensor)
                 {
-                    dev = CreateDeviceFromName(deviceRes.common.type, id, client);
-
-                    if (dev is null or default(Device))
-                    {
-                        logger.Warn($"Found not mapped device: {deviceRes.common.name} ({deviceRes.common.type})");
-                        continue;
-                    }
-
-                    if (dev is ZigbeeDevice zd)
-                        zd.AdapterWithId = deviceRes._id;
-
-                    _ = Devices.TryAdd(id, dev);
-
-                    if (!DbProvider.AddDeviceToDb(dev))
-                        _ = DbProvider.MergeDeviceWithDbData(dev);
-
-                    // If name is only numbers, try to get better name
-                    if (string.IsNullOrWhiteSpace(dev.FriendlyName) || long.TryParse(dev.FriendlyName, out _))
-                        dev.FriendlyName = deviceRes.common.name;
+                    if (ioObject.ValueName == "battery_voltage")
+                        ioObject.ValueName = "voltage";
+                    else if (ioObject.ValueName == "battery_percent")
+                        ioObject.ValueName = "battery";
                 }
-
-                foreach (var item in deviceStates.Where(x => x._id.Contains(deviceRes.native.id)))
+                else if (dev is FloaltPanel)
                 {
-                    var ioObject = new IoBrokerObject(BrokerEvent.StateChange, "", 0, item.common.name.ToLower().Replace(" ", "_"), new Parameter(item.val));
-                    if (dev is XiaomiTempSensor)
-                    {
-                        if (ioObject.ValueName == "battery_voltage")
-                            ioObject.ValueName = "voltage";
-                        else if (ioObject.ValueName == "battery_percent")
-                            ioObject.ValueName = "battery";
-                    }
-                    else if (dev is FloaltPanel)
-                    {
 
-                        if (ioObject.ValueName == "color_temperature")
-                            ioObject.ValueName = "colortemp";
-                        else if (ioObject.ValueName == "switch_state")
-                            ioObject.ValueName = "state";
-                    }
-                    if (dev is ZigbeeDevice zd)
-                        zd.SetPropFromIoBroker(ioObject, false);
+                    if (ioObject.ValueName == "color_temperature")
+                        ioObject.ValueName = "colortemp";
+                    else if (ioObject.ValueName == "switch_state")
+                        ioObject.ValueName = "state";
                 }
+                if (dev is ZigbeeDevice zd)
+                    zd.SetPropFromIoBroker(ioObject, false);
             }
+            dev.Initialized = true;
         }
+    }
 
-        protected virtual void Dispose(bool disposing)
+    protected virtual void Dispose(bool disposing)
+    {
+        if (!disposed)
         {
-            if (!disposed)
+            if (disposing)
             {
-                if (disposing)
+                foreach (var device in Devices.Values.OfType<IDisposable>())
                 {
-                    foreach (var device in Devices.Values.OfType<IDisposable>())
-                    {
-                        device?.Dispose();
-                    }
-                    Devices.Clear();
+                    device?.Dispose();
                 }
-
-                disposed = true;
+                Devices.Clear();
             }
-        }
 
-        public void Dispose()
-        {
-            // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
-            Dispose(disposing: true);
-            GC.SuppressFinalize(this);
+            disposed = true;
         }
+    }
+
+    public void Dispose()
+    {
+        // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
+        Dispose(disposing: true);
+        GC.SuppressFinalize(this);
     }
 }
