@@ -1,8 +1,9 @@
-﻿namespace AppBrokerASP.Devices.Painless
-{
-    public class Bridge : PainlessDevice
-    {
-        public Bridge(long nodeId, List<string> parameter) : base(nodeId) => ShowInApp = false;
+﻿namespace AppBrokerASP.Devices.Painless;
 
+public class Bridge : PainlessDevice
+{
+    public Bridge(long nodeId, List<string> parameter) : base(nodeId)
+    {
+        ShowInApp = false;
     }
 }
