@@ -1,14 +1,12 @@
-﻿namespace AppBrokerASP.Devices.Painless.Heater
-{
+﻿namespace AppBrokerASP.Devices.Painless.Heater;
 
-    public enum DayOfWeek : byte
-    {
-        Mon,
-        Tue,
-        Wed,
-        Thu,
-        Fri,
-        Sat,
-        Sun
-    }
+public enum DayOfWeek : byte
+{
+    Mon,
+    Tue,
+    Wed,
+    Thu,
+    Fri,
+    Sat,
+    Sun
 }
