@@ -1,13 +1,12 @@
 ﻿using SocketIOClient;
 
-namespace AppBrokerASP.Devices.Zigbee
+namespace AppBrokerASP.Devices.Zigbee;
+
+[DeviceName("lumi.router")]
+public class LumiRouter : ZigbeeDevice
 {
-    [DeviceName("lumi.router")]
-    public class LumiRouter : ZigbeeDevice
+    public LumiRouter(long nodeId, SocketIO socket) : base(nodeId, socket)
     {
-        public LumiRouter(long nodeId, SocketIO socket) : base(nodeId, socket)
-        {
-            ShowInApp = false;
-        }
+        ShowInApp = false;
     }
 }
