@@ -1,12 +1,11 @@
 ﻿using SocketIOClient;
 
-namespace AppBrokerASP.Devices.Zigbee
+namespace AppBrokerASP.Devices.Zigbee;
+
+[DeviceName("Classic B40 TW - LIGHTIFY", "AB32840")]
+public class OsramB40RW : ZigbeeLamp
 {
-    [DeviceName("Classic B40 TW - LIGHTIFY", "AB32840")]
-    public class OsramB40RW : ZigbeeLamp
+    public OsramB40RW(long nodeId, SocketIO socket) : base(nodeId, socket)
     {
-        public OsramB40RW(long nodeId, SocketIO socket) : base(nodeId, socket)
-        {
-        }
     }
 }
