@@ -1,13 +1,12 @@
 ﻿using SocketIOClient;
 
-namespace AppBrokerASP.Devices.Zigbee
+namespace AppBrokerASP.Devices.Zigbee;
+
+[DeviceName("Plug 01", "AB3257001NJ")]
+public class OsramPlug : ZigbeeSwitch
 {
-    [DeviceName("Plug 01", "AB3257001NJ")]
-    public class OsramPlug : ZigbeeSwitch
+    public OsramPlug(long nodeId, SocketIO socket) : base(nodeId, socket)
     {
-        public OsramPlug(long nodeId, SocketIO socket) : base(nodeId, socket)
-        {
-            ShowInApp = true;
-        }
+        ShowInApp = true;
     }
 }
