@@ -1,12 +1,11 @@
 ﻿using SocketIOClient;
 
-namespace AppBrokerASP.Devices.Zigbee
+namespace AppBrokerASP.Devices.Zigbee;
+
+[DeviceName("E1603/E1702")]
+public class TradfriControlOutlet : ZigbeeSwitch
 {
-    [DeviceName("E1603/E1702")]
-    public class TradfriControlOutlet : ZigbeeSwitch
+    public TradfriControlOutlet(long nodeId, SocketIO socket) : base(nodeId, socket)
     {
-        public TradfriControlOutlet(long nodeId, SocketIO socket) : base(nodeId, socket)
-        {
-        }
     }
 }
