@@ -1,4 +1,6 @@
-﻿using Newtonsoft.Json.Linq;
+﻿using AppBroker.Core;
+
+using Newtonsoft.Json.Linq;
 
 using PainlessMesh;
 
@@ -6,31 +8,31 @@ using SocketIOClient;
 
 using System.Threading.Tasks;
 
-namespace AppBrokerASP.Devices.Zigbee
+namespace AppBrokerASP.Devices.Zigbee;
+
+[AppBroker.ClassPropertyChangedAppbroker]
+public abstract partial class ZigbeeSwitch : UpdateableZigbeeDevice
 {
-    public abstract class ZigbeeSwitch : UpdateableZigbeeDevice
+    private bool state;
+
+    protected ZigbeeSwitch(long nodeId, SocketIO socket) : base(nodeId, socket)
     {
-        public bool State { get; set; }
+        ShowInApp = true;
+    }
 
-        protected ZigbeeSwitch(long nodeId, SocketIO socket) : base(nodeId, socket)
+    public override async Task UpdateFromApp(Command command, List<JToken> parameters)
+    {
+        switch (command)
         {
-            ShowInApp = true;
-        }
+            case Command.On:
+                State = true;
+                await SetValue(nameof(State), State);
+                break;
 
-        public override async Task UpdateFromApp(Command command, List<JToken> parameters)
-        {
-            switch (command)
-            {
-                case Command.On:
-                    State = true;
-                    await SetValue(nameof(State), State);
-                    break;
-
-                case Command.Off:
-                    State = false;
-                    await SetValue(nameof(State), State);
-                    break;
-            }
+            case Command.Off:
+                State = false;
+                await SetValue(nameof(State), State);
+                break;
         }
     }
 }
