@@ -1,32 +1,30 @@
-﻿namespace AppBrokerASP.IOBroker
+﻿namespace AppBrokerASP.IOBroker;
+
+public class IoBrokerHistory
 {
-    public class IoBrokerHistory
+    public IoBrokerHistory(HistoryRecord[] historyRecords, string propertyName)
     {
-        public IoBrokerHistory(HistoryRecord[] historyRecords, string propertyName)
-        {
-            HistoryRecords = historyRecords;
-            PropertyName = propertyName;
-        }
-        public IoBrokerHistory()
-        {
-            HistoryRecords = Array.Empty<HistoryRecord>();
-            PropertyName = "";
-        }
+        HistoryRecords = historyRecords;
+        PropertyName = propertyName;
+    }
+    public IoBrokerHistory()
+    {
+        HistoryRecords = Array.Empty<HistoryRecord>();
+        PropertyName = "";
+    }
 
-        public HistoryRecord[] HistoryRecords { get; set; }
-        
-        public string PropertyName { get; set; }
+    public HistoryRecord[] HistoryRecords { get; set; }
 
-        public IoBrokerHistory(string propertyName) : this()
-        {
-            PropertyName = propertyName;
-        }
+    public string PropertyName { get; set; }
 
-        public class HistoryRecord
-        {
-            public float? val { get; set; }
-            public long ts { get; set; }
-        }
+    public IoBrokerHistory(string propertyName) : this()
+    {
+        PropertyName = propertyName;
+    }
 
+    public class HistoryRecord
+    {
+        public float? val { get; set; }
+        public long ts { get; set; }
     }
 }
