@@ -2,35 +2,33 @@
 
 using System.Diagnostics;
 
-namespace AppBrokerASP.IOBroker
+namespace AppBrokerASP.IOBroker;
+
+[DebuggerDisplay("{_id}")]
+public class IoBrokerStateResponse
 {
-
-    [DebuggerDisplay("{_id}")]
-    public class IoBrokerStateResponse
+    public IoBrokerStateResponse(JToken val, bool ack, long ts, int q, string from, long lc, string type, Common common, string id, Acl acl)
     {
-        public IoBrokerStateResponse(JToken val, bool ack, long ts, int q, string from, long lc, string type, Common common, string id, Acl acl)
-        {
-            this.val = val;
-            this.ack = ack;
-            this.ts = ts;
-            this.q = q;
-            this.from = from;
-            this.lc = lc;
-            this.type = type;
-            this.common = common;
-            _id = id;
-            this.acl = acl;
-        }
-
-        public Newtonsoft.Json.Linq.JToken val { get; set; }
-        public bool ack { get; set; }
-        public long ts { get; set; }
-        public int q { get; set; }
-        public string from { get; set; }
-        public long lc { get; set; }
-        public string type { get; set; }
-        public Common common { get; set; }
-        public string _id { get; set; }
-        public Acl acl { get; set; }
+        this.val = val;
+        this.ack = ack;
+        this.ts = ts;
+        this.q = q;
+        this.from = from;
+        this.lc = lc;
+        this.type = type;
+        this.common = common;
+        _id = id;
+        this.acl = acl;
     }
+
+    public Newtonsoft.Json.Linq.JToken val { get; set; }
+    public bool ack { get; set; }
+    public long ts { get; set; }
+    public int q { get; set; }
+    public string from { get; set; }
+    public long lc { get; set; }
+    public string type { get; set; }
+    public Common common { get; set; }
+    public string _id { get; set; }
+    public Acl acl { get; set; }
 }
