@@ -5,101 +5,114 @@ using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace AppBrokerASP
+namespace AppBrokerASP;
+
+public class ServerSocket
 {
-    public class ServerSocket
+    public event EventHandler<BaseClient>? OnClientConnected;
+
+    private TcpListener? tcpListener;
+    private readonly ConcurrentDictionary<BaseClient, byte> clients;
+    private readonly ConcurrentQueue<SendMessageForQueue> sendQueue;
+    private readonly List<BaseClient> clientsToRemove = new();
+    private readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
+
+    public ServerSocket()
     {
-        public event EventHandler<BaseClient>? OnClientConnected;
+        clients = new ConcurrentDictionary<BaseClient, byte>();
+        sendQueue = new ConcurrentQueue<SendMessageForQueue>();
+        _ = Task.Run(SendMessagesFromQueue);
+    }
 
-        private TcpListener? tcpListener;
-        private readonly ConcurrentBag<BaseClient> clients;
-        private readonly Task sendTask;
-        private readonly ConcurrentQueue<SendMessageForQueue> sendQueue;
-        private readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
+    public void Start(IPAddress address, int port)
+    {
+        tcpListener = new TcpListener(address, port);
+        tcpListener.Start();
+        _ = tcpListener.BeginAcceptTcpClient(OnClientAccepted, null);
+    }
 
-        public ServerSocket()
+    public void Stop()
+    {
+        foreach (var item in clients.Keys)
+            item.Disconnect();
+        clients.Clear();
+
+        tcpListener?.Stop();
+    }
+
+    private void OnClientAccepted(IAsyncResult ar)
+    {
+        var tmpListen = tcpListener!.EndAcceptTcpClient(ar);
+        var tmpClient = new BaseClient(tmpListen);
+
+        _ = clients.TryAdd(tmpClient, 0);
+        OnClientConnected?.Invoke(this, tmpClient);
+
+        _ = tmpClient.Start();
+
+        _ = tcpListener.BeginAcceptTcpClient(OnClientAccepted, null);
+    }
+
+    public void SendToAllClients(PackageType packageType, Memory<byte> data, uint nodeId, bool logMessage = true)
+        => sendQueue.Enqueue(new SendMessageForQueue { PackageType = packageType, Data = data, NodeId = nodeId, LogMessage = logMessage });
+
+    public void SendToAllClients(PackageType packageType, Memory<byte> data, bool logMessage = true)
+        => sendQueue.Enqueue(new SendMessageForQueue { PackageType = packageType, Data = data, NodeId = 0, LogMessage = logMessage });
+
+    private void SendMessagesFromQueue()
+    {
+        while (true)
         {
-            clients = new ConcurrentBag<BaseClient>();
-            sendQueue = new ConcurrentQueue<SendMessageForQueue>();
-            sendTask = Task.Run(SendMessagesFromQueue);
+            DequeAndSend();
         }
 
-        public void Start(IPAddress address, int port)
+        void DequeAndSend()
         {
-            tcpListener = new TcpListener(new IPAddress(new byte[] { 0, 0, 0, 0 }), port);
-            tcpListener.Start();
-            _ = tcpListener.BeginAcceptTcpClient(OnClientAccepted, null);
-        }
-
-        public void Stop()
-        {
-            foreach (var item in clients)
-                item.Disconnect();
-            clients.Clear();
-
-            tcpListener?.Stop();
-        }
-
-
-        private void OnClientAccepted(IAsyncResult ar)
-        {
-            var tmpListen = tcpListener!.EndAcceptTcpClient(ar);
-            var tmpClient = new BaseClient(tmpListen);
-            clients.Add(tmpClient);
-            OnClientConnected?.Invoke(this, tmpClient);
-
-            _ = tmpClient.Start();
-
-            _ = tcpListener.BeginAcceptTcpClient(OnClientAccepted, null);
-        }
-
-        public void SendToAllClients(PackageType packageType, Memory<byte> data, uint nodeId, bool logMessage = true)
-            => sendQueue.Enqueue(new SendMessageForQueue { PackageType = packageType, Data = data, NodeId = nodeId, LogMessage = logMessage });
-
-        public void SendToAllClients(PackageType packageType, Memory<byte> data, bool logMessage = true)
-            => sendQueue.Enqueue(new SendMessageForQueue { PackageType = packageType, Data = data, NodeId = 0, LogMessage = logMessage });
-
-        private void SendMessagesFromQueue()
-        {
-            while (true)
+            if (sendQueue.TryDequeue(out var msg))
             {
-                if (sendQueue.TryDequeue(out var msg))
+                foreach (var keyClient in clients)
                 {
-                    foreach (var client in clients)
+                    var client = keyClient.Key;
+                    try
                     {
-                        try
+                        if (client.Source.IsCancellationRequested)
                         {
-                            if (client.Source.IsCancellationRequested)
-                            {
-                                _ = clients.TryTake(out var a);
-                                continue;
-                            }
-                            //if (msg.LogMessage)
-                            //    logger.Debug($"Send to NodeId: {msg.NodeId}, Type: {msg.PackageType}, Data: {msg.Data}");
-                            client.Send(msg.PackageType, msg.Data.Span, msg.NodeId);
+                            clientsToRemove.Add(client);
+                            continue;
                         }
-                        catch (Exception e)
-                        {
-                            Console.WriteLine(e);
-                            logger.Error(e);
-                            _ = clients.TryTake(out var a);
-                        }
+                        //if (msg.LogMessage)
+                        //    logger.Debug($"Send to NodeId: {msg.NodeId}, Type: {msg.PackageType}, Data: {msg.Data}");
+                        client.Send(msg.PackageType, msg.Data.Span, msg.NodeId);
                     }
-                    Thread.Sleep(100);
+                    catch (Exception e)
+                    {
+                        Console.WriteLine(e);
+                        logger.Error(e);
+                        clientsToRemove.Add(client);
+                    }
                 }
-                else
+                if (clientsToRemove.Count > 0)
                 {
-                    Thread.Sleep(16);
+                    foreach (var item in clientsToRemove)
+                    {
+                        _ = clients.Remove(item, out _);
+                    }
+                    clientsToRemove.Clear();
                 }
+                Thread.Sleep(100);
+            }
+            else
+            {
+                Thread.Sleep(16);
             }
         }
+    }
 
-        private class SendMessageForQueue
-        {
-            public PackageType PackageType { get; set; }
-            public Memory<byte> Data { get; set; }
-            public uint NodeId { get; set; }
-            public bool LogMessage { get; internal set; }
-        }
+    private class SendMessageForQueue
+    {
+        public PackageType PackageType { get; set; }
+        public Memory<byte> Data { get; set; }
+        public uint NodeId { get; set; }
+        public bool LogMessage { get; internal set; }
     }
 }
