@@ -1,12 +1,9 @@
-﻿using System;
+﻿
+using System;
 
-namespace DbMigrator
+namespace DbMigrator;
+
+class Program
 {
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
+    static void Main(string[] args) => Console.WriteLine("Hello World!");
 }
