@@ -1,23 +1,21 @@
-﻿using Newtonsoft.Json;
+﻿using AppBroker.Core;
+
+using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 using System.Collections;
 using System.Text;
 
-namespace PainlessMesh
+namespace PainlessMesh;
+
+public abstract class BaseSmarthomeMessage
 {
+    [JsonProperty("id")]
+    public virtual uint NodeId { get; set; }
 
-    public abstract class BaseSmarthomeMessage
-    {
-        [JsonProperty("id")]
-        public virtual uint NodeId { get; set; }
+    [JsonProperty("m"), JsonConverter(typeof(StringEnumConverter))]
+    public virtual MessageType MessageType { get; set; }
 
-        [JsonProperty("m"), JsonConverter(typeof(StringEnumConverter))]
-        public virtual MessageType MessageType { get; set; }
-
-        [JsonProperty("c"), JsonConverter(typeof(StringEnumConverter))]
-        public virtual Command Command { get; set; }
-
-    }
-
+    [JsonProperty("c"), JsonConverter(typeof(StringEnumConverter))]
+    public virtual Command Command { get; set; }
 }
