@@ -1,4 +1,6 @@
-﻿using Newtonsoft.Json;
+﻿using AppBroker.Core;
+
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
 using System.Collections.Generic;
@@ -7,30 +9,29 @@ using System.Linq;
 
 using nj = Newtonsoft.Json;
 
-namespace PainlessMesh
+namespace PainlessMesh;
+
+//[NonSucking.Framework.Serialization.Nooson]
+public partial class JsonSmarthomeMessage : BaseSmarthomeMessage
 {
-    public partial class JsonSmarthomeMessage : BaseSmarthomeMessage
+    [nj.JsonProperty("p")]
+    public List<JToken> Parameters { get; set; }
+    [JsonIgnore]
+    public override uint NodeId { get; set; }
+
+    [JsonProperty("id")]
+    public long LongNodeId { get; set; }
+
+    public JsonSmarthomeMessage(uint nodeId, MessageType messageType, Command command, params JToken[] parameters)
     {
-        [nj.JsonProperty("p")]
-        public List<JToken> Parameters { get; set; }
-        [JsonIgnore]
-        public override uint NodeId { get; set; }
-
-        [JsonProperty("id")]
-        public long LongNodeId { get; set; }
-
-        public JsonSmarthomeMessage(uint nodeId, MessageType messageType, Command command, params JToken[] parameters)
-        {
-            NodeId = nodeId;
-            MessageType = messageType;
-            Command = command;
-            Parameters = parameters.ToList();
-        }
-
-        public JsonSmarthomeMessage()
-        {
-
-        }
+        NodeId = nodeId;
+        MessageType = messageType;
+        Command = command;
+        Parameters = parameters.ToList();
     }
 
+    public JsonSmarthomeMessage()
+    {
+
+    }
 }
