@@ -1,20 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace PainlessMesh.Ota
+namespace PainlessMesh.Ota;
+
+
+
+internal partial struct FirmwareAdvertisment
 {
-    internal struct FirmwareAdvertisment
+    public List<byte[]> Parts { get; init; }
+    public DateTime AdvertiseUntil { get; init; }
+    public FirmwareMetadata Metadata { get; init; }
+
+    public FirmwareAdvertisment(List<byte[]> parts, DateTime advertiseUntil, FirmwareMetadata metadata)
     {
-        public List<byte[]> Parts { get; init; }
-        public DateTime AdvertiseUntil { get; init; }
-        public FirmwareMetadata Metadata { get; init; }
-
-
-        public FirmwareAdvertisment(List<byte[]> parts, DateTime advertiseUntil, FirmwareMetadata metadata)
-        {
-            Parts = parts;
-            AdvertiseUntil = advertiseUntil;
-            Metadata = metadata;
-        }
+        Parts = parts;
+        AdvertiseUntil = advertiseUntil;
+        Metadata = metadata;
     }
 }
