@@ -1,12 +1,8 @@
-﻿namespace PainlessMesh
+﻿namespace PainlessMesh;
+
+public enum SmarthomePackageType : byte
 {
-
-
-    public enum SmarthomePackageType : byte
-    {
-        Error = 0,
-        Normal = 1,
-        Ota = 123
-    }
-
+    Error = 0,
+    Normal = 1,
+    Ota = 123
 }
