@@ -1,13 +1,7 @@
 ﻿using Microsoft.CodeAnalysis.Text;
 using Microsoft.CodeAnalysis;
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.Diagnostics;
-using System.Collections.Immutable;
 
 namespace AppBroker.Generators
 {
@@ -29,6 +23,22 @@ namespace AppBroker
 }
 ";
 
+
+        private const string addOverrideAttribute = @"
+using System;
+namespace AppBroker
+{
+    [AttributeUsage(AttributeTargets.Field, Inherited = false, AllowMultiple = false)]
+    [System.Diagnostics.Conditional(""NotifyPropertyChangedGenerator_DEBUG"")]
+    sealed class AddOverrideAttribute : Attribute
+    {
+        public AddOverrideAttribute()
+        {
+        }
+    }
+}
+";
+
         private const string ignoreChangedFieldAttribute = @"
 using System;
 namespace AppBroker
@@ -40,24 +50,6 @@ namespace AppBroker
         public IgnoreChangedFieldAttribute()
         {
         }
-    }
-}
-";
-
-        private const string copyPropertyAttributesFromAttribute = @"
-using System;
-namespace AppBroker
-{
-    [AttributeUsage(AttributeTargets.Field, Inherited = false, AllowMultiple = false)]
-    [System.Diagnostics.Conditional(""NotifyPropertyChangedGenerator_DEBUG"")]
-    sealed class CopyPropertyAttributesFromAttribute : Attribute
-    {
-        public CopyPropertyAttributesFromAttribute(string propertyName)
-        {
-            PropertyName = propertyName;
-        }
-
-        public string PropertyName { get; set; }
     }
 }
 ";
@@ -116,8 +108,8 @@ namespace AppBroker
                 i.AddSource("IgnoreChangedFieldAttribute", ignoreChangedFieldAttribute);
                 i.AddSource("ClassPropertyChangedAppbrokerAttribute", classAttribute);
                 i.AddSource("PropertyChangedAppbrokerAttribute", propertyChangedFieldAttribute);
-                i.AddSource("CopyPropertyAttributesFromAttribute", copyPropertyAttributesFromAttribute);
                 i.AddSource("IgnoreFieldAttribute", ignoreFieldAttribute);
+                i.AddSource("AddOverrideAttribute", addOverrideAttribute);
             });
 
             // Register a syntax receiver that will be created for each generation pass
@@ -127,22 +119,34 @@ namespace AppBroker
         public void Execute(GeneratorExecutionContext context)
         {
             // retrieve the populated receiver 
-            if (!(context.SyntaxContextReceiver is SyntaxReceiver receiver))
+            if (context.SyntaxContextReceiver is not SyntaxReceiver receiver)
                 return;
+
+            //Debugger.Launch();
 
             // get the added attribute, and INotifyPropertyChanged
             INamedTypeSymbol attributeSymbol = context.Compilation.GetTypeByMetadataName("AppBroker.PropertyChangedAppbrokerAttribute");
             INamedTypeSymbol ignoreChangedFieldSymbol = context.Compilation.GetTypeByMetadataName("AppBroker.IgnoreChangedFieldAttribute");
-            INamedTypeSymbol ignoreFieldSymbol= context.Compilation.GetTypeByMetadataName("AppBroker.IgnoreFieldAttribute");
+            INamedTypeSymbol ignoreFieldSymbol = context.Compilation.GetTypeByMetadataName("AppBroker.IgnoreFieldAttribute");
+            INamedTypeSymbol overrideFieldSymbol = context.Compilation.GetTypeByMetadataName("AppBroker.AddOverrideAttribute");
 
             foreach (var group in receiver.Classes)
             {
-                string classSource = ProcessClass(group.Key as INamedTypeSymbol, group.Value, attributeSymbol, ignoreChangedFieldSymbol, ignoreFieldSymbol);
-                context.AddSource($"{group.Key.Name}_autoNotifyAppbroker.cs", SourceText.From(classSource, Encoding.UTF8));
+                string classSource = ProcessClass(group.Key as INamedTypeSymbol, group.Value, attributeSymbol, ignoreChangedFieldSymbol, ignoreFieldSymbol, overrideFieldSymbol, context);
+                context.AddSource($"{group.Key.Name}.Appbroker.cs", SourceText.From(classSource, Encoding.UTF8));
             }
         }
 
-        private string ProcessClass(INamedTypeSymbol classSymbol, SyntaxReceiver.ClassThingy thingy, INamedTypeSymbol attributeSymbol, INamedTypeSymbol ignoreAttributeSymbol, INamedTypeSymbol ignoreFieldSymbol)
+        private bool BaseClassAlreadyHasAttribute(INamedTypeSymbol parent, string attributeName)
+        {
+            if (parent.BaseType == null)
+                return false;
+            var attr = parent.BaseType.GetAttributes();
+
+            return attr.Any(x => $"{x.AttributeClass.ContainingNamespace}.{x.AttributeClass.Name}" == attributeName) || BaseClassAlreadyHasAttribute(parent.BaseType, attributeName);
+        }
+
+        private string ProcessClass(INamedTypeSymbol classSymbol, SyntaxReceiver.ClassThingy thingy, INamedTypeSymbol attributeSymbol, INamedTypeSymbol ignoreAttributeSymbol, INamedTypeSymbol ignoreFieldSymbol, INamedTypeSymbol overrideFieldSymbol, GeneratorExecutionContext context)
         {
             if (!classSymbol.ContainingSymbol.Equals(classSymbol.ContainingNamespace, SymbolEqualityComparer.Default))
             {
@@ -151,17 +155,23 @@ namespace AppBroker
 
             string namespaceName = classSymbol.ContainingNamespace.ToDisplayString();
 
+            var alreadyContainsMethod = BaseClassAlreadyHasAttribute(classSymbol, "AppBroker.ClassPropertyChangedAppbrokerAttribute");
+            string additional = "";
+            if (classSymbol.IsGenericType)
+                additional = $"<{string.Join(",", classSymbol.TypeParameters.Select(x => x.Name))}>";
+
             // begin building the generated source
-            StringBuilder source = new StringBuilder($@"
+            StringBuilder source = new ($@"
 using System.Runtime.CompilerServices;
 namespace {namespaceName}
 {{
-    public partial class {classSymbol.Name}");
+    public partial class {classSymbol.Name}{additional}");
 
             // if the class doesn't implement INotifyPropertyChanged already, add it
 
-            source.Append($@"
+            _ = source.Append($@"
     {{
+{(alreadyContainsMethod ? "" : $@"
         protected T RaiseAndSetIfChanged<T>(ref T field, T value, [CallerMemberName] string? propertyName = null)
         {{
             if (System.Collections.Generic.EqualityComparer<T>.Default.Equals(field, value))
@@ -171,29 +181,22 @@ namespace {namespaceName}
             {(thingy.CallPropertyChanged ? "OnPropertyChanged(propertyName);" : "")}
             return value;
         }}
+")}
             ");
-
-
-
-            //foreach (IPropertySymbol propSymbol in thingy.Properties)
-            //{
-            //    ProcessProp(source, propSymbol);
-            //}
 
             // create properties for each field 
             foreach (IFieldSymbol fieldSymbol in thingy.Fields)
             {
-                thingy.AdditionalAttributesForGeneratedProp.TryGetValue(fieldSymbol, out var additionalAttributes);
+                _ = thingy.AdditionalAttributesForGeneratedProp.TryGetValue(fieldSymbol, out var additionalAttributes);
 
-                ProcessField(source, fieldSymbol, attributeSymbol, ignoreAttributeSymbol, ignoreFieldSymbol, additionalAttributes);
+                ProcessField(source, fieldSymbol, attributeSymbol, ignoreAttributeSymbol, ignoreFieldSymbol,overrideFieldSymbol, additionalAttributes, context);
             }
 
-            source.Append(" }\n}");
+            _ = source.Append("    }\n}");
             return source.ToString();
         }
 
-
-        private void ProcessField(StringBuilder source, IFieldSymbol fieldSymbol, ISymbol attributeSymbol, INamedTypeSymbol ignoreAttributeSymbol, INamedTypeSymbol ignoreFieldSymbol, ImmutableArray<AttributeData> additionalAttributes)
+        private void ProcessField(StringBuilder source, IFieldSymbol fieldSymbol, ISymbol attributeSymbol, INamedTypeSymbol ignoreAttributeSymbol,INamedTypeSymbol ignoreFieldSymbol, INamedTypeSymbol overrideFieldSymbol, List<AttributeListSyntax> additionalAttributes, GeneratorExecutionContext context)
         {
             // get the name and type of the field
             string fieldName = fieldSymbol.Name;
@@ -215,20 +218,50 @@ namespace {namespaceName}
             {
                 //TODO: issue a diagnostic that we can't process this field
                 return;
+
+    //            NoosonGenerator.MakeDiagnostic("0005",
+    //"",
+    //"IEnumerable is not supported for deserialization, implement own deserializer or this value will be lost.",
+    //property.Symbol,
+    //DiagnosticSeverity.Error
+    //);
             }
 
-            if (additionalAttributes != null && additionalAttributes.Length > 0)
+            if (additionalAttributes != null && additionalAttributes.Count > 0)
+            {
                 foreach (var additionalAttribute in additionalAttributes)
                 {
-                    if(additionalAttribute.ApplicationSyntaxReference.GetSyntax() is AttributeSyntax ass)
+                    var mod = context.Compilation.GetSemanticModel(additionalAttribute.SyntaxTree);
+                    foreach (var addAttr in additionalAttribute.Attributes)
                     {
-                        source.AppendLine($"        [{additionalAttribute.AttributeClass.ToDisplayString()}{ass.ArgumentList.ToFullString()}]");
+                        var symb = mod.GetSymbolInfo(addAttr);
+                        var candidate = symb.Symbol ?? symb.CandidateSymbols.FirstOrDefault();
+                        if (candidate == null)
+                        {
+                            context.ReportDiagnostic(
+                                Diagnostic.Create(
+                                    "AB0234", 
+                                    "Generator",
+                                    $"The type or namespace name '{addAttr.Name}' does not exist.' (did you use a the wrong namespace while declaring the fullname of the property attribute?)", 
+                                    DiagnosticSeverity.Error, 
+                                    DiagnosticSeverity.Error, 
+                                    true, 
+                                    0, 
+                                    location: Location.Create(addAttr.SyntaxTree, addAttr.Span)));
+                            _ = source.AppendLine($"        [{addAttr.ToFullString()}]");
+                        }
+                        else
+                        {
+                            _ = source.AppendLine($"        [{candidate.ContainingType.ToDisplayString()}{addAttr.ArgumentList.ToFullString()}]");
+                        }
                     }
-
                 }
+            }
 
-            source.Append($@"
-        public {fieldType} {propertyName} 
+            var @override = fieldAttrs.Any(ad => ad.AttributeClass.Equals(overrideFieldSymbol, SymbolEqualityComparer.Default));
+
+            _ = source.Append($@"
+        public {(@override ? "override " : "")}{fieldType} {propertyName} 
         {{
             get => this.{fieldName};
             set => {(ignore ? $"this.{fieldName} = value;" : $"this.RaiseAndSetIfChanged(ref {fieldName}, value);")}
@@ -246,14 +279,9 @@ namespace {namespaceName}
                 if (fieldName.Length == 0)
                     return string.Empty;
 
-                if (fieldName.Length == 1)
-                    return fieldName.ToUpper();
-
-                return fieldName.Substring(0, 1).ToUpper() + fieldName.Substring(1);
+                return fieldName.Length == 1 ? fieldName.ToUpper() : fieldName.Substring(0, 1).ToUpper() + fieldName.Substring(1);
             }
-
         }
-
 
         /// <summary>
         /// Created on demand before each generation pass
@@ -270,11 +298,9 @@ namespace {namespaceName}
                 public bool CallPropertyChanged { get; set; }
                 public bool CallPropertyChanging { get; set; }
 
-
-
                 public List<IFieldSymbol> Fields { get; } = new List<IFieldSymbol>();
 
-                public Dictionary<IFieldSymbol, System.Collections.Immutable.ImmutableArray<AttributeData>> AdditionalAttributesForGeneratedProp = new();
+                public Dictionary<IFieldSymbol, List<AttributeListSyntax>> AdditionalAttributesForGeneratedProp = new();
 
             }
             public Dictionary<ISymbol, ClassThingy> Classes { get; } = new();
@@ -290,13 +316,12 @@ namespace {namespaceName}
                     foreach (VariableDeclaratorSyntax variable in fieldDeclarationSyntax.Declaration.Variables)
                     {
                         // Get the symbol being declared by the field, and keep it if its annotated
-                        IFieldSymbol fieldSymbol = context.SemanticModel.GetDeclaredSymbol(variable) as IFieldSymbol;
+                        var fieldSymbol = context.SemanticModel.GetDeclaredSymbol(variable) as IFieldSymbol;
                         if (fieldSymbol.GetAttributes().Any(ad => ad.AttributeClass.ToDisplayString() == "AppBroker.PropertyChangedAppbrokerAttribute"))
                         {
                             var symbol = fieldSymbol.ContainingType as ISymbol;
                             if (!Classes.TryGetValue(symbol, out var thingy))
                                 Classes[symbol] = thingy = new ClassThingy(symbol);
-
 
                             if (thingy.Fields.Contains(fieldSymbol))
                                 continue;
@@ -331,7 +356,6 @@ namespace {namespaceName}
                             IParameterSymbol item = symbolAttr.AttributeConstructor.Parameters[i];
                             var parameterValue = symbolAttr.ConstructorArguments[i].Value;
 
-
                             switch (item.Name)
                             {
                                 case "callPropertyChanged":
@@ -352,31 +376,33 @@ namespace {namespaceName}
                             if (p is IFieldSymbol fieldSymbol && !thingy.Fields.Contains(fieldSymbol))
                             {
                                 thingy.Fields.Add(fieldSymbol);
-                                var fieldAttributes = fieldSymbol.GetAttributes();
-                                if (fieldAttributes.Length == 0)
-                                    continue;
-                                var copyFromAttribute = fieldAttributes.FirstOrDefault(x => x.AttributeClass.Name == "CopyPropertyAttributesFromAttribute");
 
-                                if (copyFromAttribute == default)
+                                var declaringSyntax 
+                                    = fieldSymbol
+                                    .DeclaringSyntaxReferences
+                                    .FirstOrDefault();
+
+                                if (declaringSyntax == default)
                                     continue;
 
-                                var propName = copyFromAttribute.ConstructorArguments.First();
-                                var propertyMember = allMembers.FirstOrDefault(x => x.Name == propName.Value.ToString());
+                                var attribuesForProp
+                                    = declaringSyntax
+                                    .GetSyntax()
+                                    .Parent
+                                    .Parent
+                                    .ChildNodes()
+                                    .OfType<AttributeListSyntax>()
+                                    .Where(x => x.Target?.Identifier.Text == "property")
+                                    .ToList();
 
-                                if (propertyMember is not IPropertySymbol propertySymbol)
-                                {
-                                    //TODO: Issue Warning for wrong use of attribute
+                                if (attribuesForProp.Count == 0)
                                     continue;
-                                }
-                                var attributeCopy = propertySymbol.GetAttributes();
-                                thingy.AdditionalAttributesForGeneratedProp[fieldSymbol] = attributeCopy;
+
+                                thingy.AdditionalAttributesForGeneratedProp[fieldSymbol] = attribuesForProp;
                             }
-
-
                         }
                     }
                 }
-
             }
         }
     }
