@@ -4,10 +4,12 @@
     {
         public static void Main(string[] args)
         {
-            var asd = new PropChanged();
-            asd.MyProperty = 123;
-            asd.MyProperty2 = 123;
-            asd.Different = 123;
+            var asd = new PropChanged
+            {
+                MyProperty = 123,
+                MyProperty2 = 123
+            };
+            //asd.Different = 123;
         }
     }
 }
